requirements.txt
aiogram
python-dotenv
edge-tts
Pillow
aiohttp
pilmoji